bdr: Initial implementation of a more elaborate tuple wire protocol.
BDR wants to be fast, but it also wants to support replication across
versions and architectures. So far we used to always copy the binary
data as is between sender as receiver. That works well enough for
simple scalar core data types, but it breaks down when oids are
embedded in the disk format.

Instead allow to specify the format on a per-datum format. It's now
supported to send data in binary, send/recv format or text. On the
sending side we select binary if it's a scalar core type, and
otherwise use send/recv. This needs to be extended to support cross
architecture replication, but that's for another commit.