bdr: Use dynamic background workers and shared memory
BDR previously used one static bgworker for each apply connection, passing the
configuration for each apply worker as a pointer into postmaster memory.

This relied on the ability to fork() without exec(), which won't work on
EXEC_BACKEND platforms (i.e. Windows). It also prevented any possibility
of adding new connections at runtime, made it harder to control when
apply workers started, and made it impossible to run a short lived apply
worker.

In order to properly support dump/apply for logical dump based node bring-up,
we need to be able to run short-lived bgworkers and pass parameters to them
that are only determined after postmaster start. This requires dynamic
bgworkers and the use of shared memory to communicate parameters to bgworkers.