deparse/core: have ALTER TABLE return table OID and other data
Currently, most if not all CREATE commands return the OID of the created
object.  This is useful for event triggers to try to figure out exactly
what happened.  This patch extends this idea a bit further to cover
ALTER TABLE as well: auxilliary info such as a second OID or a column
number are also returned for each subcommand of the ALTER TABLE.  This
OID or attnum can later be used to determine, for instance, which
constraint was added by ALTER TABLE ADD CONSTRAINT, or which parent got
dropped from the inherit-from parents list by NO INHERIT.  This makes it
possible to decode with precision exactly what happened during execution
of any ALTER TABLE command.

There is no user-visible change here; this patch makes the info
available for later.