bdr: queue object drops more comprehensively
We were skipping listing objects in the queued item that are only
reached by internal dependency links, but this doesn't work in some
cases, e.g. a table that depends on a view through column dependencies.

We're probably going to need to tweak this again in a later stage, for
example if we want local-only objects (such as unlogged matviews) to
depend on replicated objects; if we were to try that scenario today, the
drop would fail because the matview wouldn't be found on the remote node
when we tried to replicate the drop.  But this implementation doesn't
seem to be the way forward for that, anyway.