bdr: Better error message on slot creation for db where bdr not active
We want to block slot creation on a database where there's no active
bdr connection configured for that db. Test to see whether there's a
bdr schema and refuse to create a slot in this case.

Doesn't protect against bdr being enabled then disabled, but nothing
exciting will happen in that case anyway.