bdr: Add db oid param to fetch_sysid_via_node_id, add bdr_ prefix
Rename fetch_sysid_via_node_id to bdr_fetch_sysid_via_node_id and add a new
output parameter to return the database oid for the database that's local
to that RepNodeId.

This allows bdr_fetch_sysid_via_node_id to return the full identifying
key for a node now that we support intra-node replication.