bdr-drops: Set "normal" flag when DEPFLAG_REVERSE
This seems to fix the problem for views depending on tables without
causing problems elsewhere.

Whether this approach is really correct remains to be seen.  Do we need
a separate flag in pg_event_trigger_get_dropped_objects() to return the
"reverse internal" status?  Seems awkward, so for now just turn it into
a "normal" dependency instead.