bdr: Significantly improve wal flushing communication.
The feedback behaviour of apply processes with senders was far from
optimal: When bdr.synchronous_commit was set to off we reported LSNs
as flushed that weren't. And without bdr.synchronous_commit=off replay
was slow unless fsync()s are fast. That's because there's no
parallelism during apply.

Improve reporting by keeping a in-memory list associating remote and
local LSNs so we can check up to where we've already flushed.

Also only accept a boolean for bdr.synchronous_commit=on/off.