bdr: don't use a global var after clobbering it
This was reported as #37825 whereby a sequence is not dropped after
creating it for a serial column; but the problem was much more general
than that.  Essentially, any time a table was created, the rest of the
command list was ignored because creating the truncate trigger reset
SPI_nprocessed, causing the outer loop to exit early.

Fix by using a local copy of the global variable.  Also, add the
originally reported problem as a test case in ddl/create.sql.