bdr: Don't print update conflicts using the indexes relation descriptor.
For a long while now we're using the heap's and not the indexes
relation descriptor. But apparently do_log_update() didn't get the
memo.
That happened to work fine for all our testcases because the indexes
and the heap definition were sufficiently compatible - but it's far
from guaranteed to work.

Per report from Keaton Adams.