bdr: Initial setup of isolation tester for bdr
At this point this only adds the Makefile infrastructure, config file, and a
first script to make sure the cluster is up and ready.

It uses DBs named node1, node2 and node3, and does an init_replica from node1
to the other two.

It's causing memory corruption, so that needs to be figured out before things
can proceed.