bdr: Move the pg_sleep() into a separate "test" during isolation start
isolationtester sends the setup block as a multistatement, which causes
confusion if we need to sleep on all nodes to let DDL locking get ready, but
the statement after the sleep is DDL.

Do a fake "test" that sleeps on all nodes first, instead.

This should not be necessary. It is a workaround for startup races.