bdr: Don't attempt to use get_database_name unsafely
The prior commit "bdr: Show database name when a DDL lock error is encountered"
added use of get_database_name(..) to clarify an error message. However, it's not
always safe to access the catalogs from this call site.

Remove the unsafe call and document the issue in the code so it won't be added
back later.