bdr: compiles and works with vanilla 9.4
This is UDR v0. Basic master->slave replication works.

Currently the build has to be done using `make UDR=1` as there is no
detection if the plugin is being built with 9.4 or the 9.4bdr.

There are still warnings (about unused variables because of #ifdefs),
bdr_init_replica and bdr_init_copy don't work yet.