deparse: infrastructure needed for command deparsing
This patch introduces unused infrastructure for command deparsing.
There are three main parts:

1. A list (stash) of executed commands in Node format, stored in the
event trigger state.  At ddl_command_end, the stored items can be
extracted.  For now this only support "basic" commands (in particular
not ALTER TABLE or GRANT).  It's useful enough to cover all CREATE
command as well as many simple ALTER forms.

2. Support code to enable writing routines that convert the Node format
into a JSON blob.  This JSON representation allows extracting and
modifying individual parts of the command.

3. Code to convert the JSON blobs back into plain strings.

No actual routines to convert Node into JSON is provided by this patch;
that is left to later patches.  This split is only presented as is for
ease of review.