replication_identifiers: Change the GetReplicationInfoByIdentifier nterface.
Previously, the GetReplicationInfoByIdentifier function returned
SysCache HeapTuple which meant that callers had to release syscache
which is not very good interface separation. The change is to make it
return the actual values from the tuple as out variables (currently
only riname is returned as that's the only one used by callers).

Note: this is a breaking change.