bdr: Launch apply workers from in-db config
Apply workers are now launched based on configuration stored in the
database, rather than configuration inherited from the postmaster
at fork() time. Each apply worker accesses the database to get its
configuration rather than using the postmaster's memory.

At this point the config file is still read and used by temporary
hooks in early per-db worker startup to populate the connection
configuration tables, but after that it is unused.

The next step is to eliminate the configuration file, making the
contents of the configuration tables in the database authorative
and persistent.