Add a node join process on top of dynamic configuration
***This commit breaks bdr_init_copy***

Changes:

The meaning of bdr_connections is changed dramatically.

It used to contain one entry for each connection from one node to another (i.e.
n^2 entries for n nodes), with the dsn identifying the data the apply worker
uses to connect to the destination. The sysid/timeline/dboid identified
the node that would connect to the dsn.

Now it has one entry per node, with the dsn pointing to that node. The
sysid/timeline/dboid identify the node to which this dsn refers.

bdr.bdr_connection_add(...) is removed.

It is now necessary to call the new bdr.node_join(...) on the root node before
it can accept any connections from other nodes. This provides the connection
information for other nodes to use.

To join new nodes to the root node, bdr.node_join(...) is used with the
init_from_dsn parameter set to a libpq connection string specifying a
connection to an existing BDR node (not necessarily the root node).

The basic logic of init remains the same. A per-db worker is started
on the joining node. This worker makes a libpq connection to the
init target node and creates a slot on it. It takes a note of the
snapshot exported by the slot. The snapshot is used to dump the remote
node; this dump is then applied to the local node's blank database.
Slots are then created on all other nodes. A temporary apply worker is ued to
stream changes in 'catchup mode' from the target node, where changes from all
nodes, not just the immediate origin, get relayed.

Once catchup is completed, a row is inserted into bdr_connections on the
target, then a function is called on the target (and inserted into the
command queue for other nodes to run) to tell other nodes to make the
inbound slots.

The perdb worker on the joining node waits for all these slots to appear
on the local node before marking the local node as ready to accept
commands.

Significant further work is required to turn this into a safe node
part/join system. In particular:

- No attempt is made to ensure that DDL global locking is not currently
  in progress or begun during the join.

- No effort is made to ensure that in-progress sequence voting is
  handled.

- There is no process for removing a node