bdr: Add caching layer above databases.
Currently that layer is just used to cache (and abstract away) checks
whether a database is enabled for bdr. But likely we'll end up caching
additional data in there.

FIXME: Right now this isn't going to work for UDR. We need to make it
set a flag on the source side for it.