Have RENAME routines return ObjectAddress rather than OID
This lets them include an objectSubId when appropriate (i.e. when
renaming columns), and additionally they make the containing catalog OID
available for further processing.  The object OID that was previously
returned is still available in the objectId field of the returned
ObjectAddress.

This isn't terribly exciting in itself but it supports future event
trigger changes.

Discussion: 20150218213255.GC6717@tamriel.snowman.net
Reviewed-By: Andres Freund, Stephen Frost