Make CREATE OR REPLACE VIEW more consistent internally
The way that columns are added to a view is by calling
AlterTableInternal with special subtype AT_AddColumnToView; but that
subtype AT_AddColumnToView is changed to AT_AddColumnRecurse by
ATPrepAddColumn.  This has no visible effect in the current code, since
views cannot have inheritance children and adding a column to a table or
a view is executed identically; but it does make a difference for event
triggers keeping track of commands, because that leads to confusing the
case with a normal ALTER TABLE ADD COLUMN.

Fix the problem by passing a flag to ATPrepAddColumn to prevent it from
changing the subtype.  The event trigger code can then properly ignore
the subcommand.