Make bdr_replicate_ddl_command() obey the command filter.
Previously, bdr_replicate_ddl_command() would always execute any
command, no matter if it's considered safe or not. But this behaviour is
too much of a footgun for users of this function. Plus we already have
mechanism to override command filter - the
bdr.permit_unsafe_ddl_commands GUC, so it makes sense that
bdr_replicate_ddl_command() function should behave according to this GUC
just like plain DDL in BDR.