Don't dump setval(...) for 'bdr' sequences
BDR sequence state is stored in the BDR extension data. It is neither
necessary nor - currently - supported to setval(...) a 'bdr' sequence.
Attempting to do so in dumps meant we had to ignore errors on restore.

Note that this fix isn't perfect. We CREATE SEQUENCE ... USING 'bdr',
which will start sequence voting immediately unless
--single-transaction is used. If pg_restore or a dump script executed
by 'psql' concurrently creates extension data this may result in
unique violations. The sequence should only be created after we've
restored bdr.bdr_sequence_values but pg_dump doesn't offer a way for a
sequence (which is considered schema) to depend on table data. This
issue is not addressed in this commit, but will only arise if you are
restoring a dump to an empty but active BDR system, rather then to a
standalone node before promotion to a BDR master.

(A possible workaround would be to CREATE the sequence as local, then
ALTER SEQUENCE ... USING 'bdr' in dumpSequenceData, but this means
that schema-only dumps will lose knowledge of the sequence access
method.)