On second thought use session cache for bdr_nodes.
This is how rest of postgres does it. The shmem cache is probably
slightly better performance-wise but it's somewhat abusing the shmem.

The node access API could use some rethinking after this - standard
access should probably go via cache. We should also avoid SPI probably.
But both of these are left for future patches.

In passing activate the read_only regression test which was previously
not added to Makefile by mistake.