Reject attempts to part from self, detect part from absent node
Reject attempts to part a node from its self. This seems like
a sensible thing to do, but we need to remove a node using another
node so that all peers find out about the removal.

Detect when asked to part a nonexistent node and complain instead
of silently doing nothing.

Also warn when asked to part a node that isn't in 'r' state.