Remove the last DO block from the 0.8.0.1 -> 0.8.0.2 upgrade script
This results in the following mysterious error:

    ERROR:  trigger "bdr_truncate_trigger_add": unrecognized event name ""
    CONTEXT:  SQL statement "CREATE EVENT TRIGGER bdr_truncate_trigger_add
                    ON ddl_command_end
                    EXECUTE PROCEDURE bdr.bdr_truncate_trigger_add()"
    PL/pgSQL function inline_code_block line 23 at SQL statement

Looking at the error as reported in the log file, the apparently empty
'unrecognized event name' is actually reported as a bunch of 0x7F
characters.

The affected DO-block executes correctly in later versions (we're
dealing with 9.4b2 here) so I presume whatever causes it has been
fixed.

For our purposes we should be able to assume the functions/triggers
covered in this block don't exist, and can work with the CREATE
statements as-is as a workaround.