Allow bdr_get_remote_nodeinfo_internal to work with 0.7.x
bdr_get_remote_nodeinfo_internal tried to call functions that
don't exist on a remote 0.7.x node. Tweak it to make it more
tolerant of version differences.

This isn't nearly enough to let a 0.9.x node join a 0.7.x
group, but it fixes the first issue that arises when doing so.