make dml/missing_pk test independent of current user
Spelling out the SQL is way more verbose than using psql meta
commands, but this way we avoid having the current user name
in the output. This unbreaks the test in case it's not running
under the name of the original author (e.g. in jenkins or
some random developer).