Hotfix for handling table rewrites
A full table rewrite emits writes to pg_temp_.... relations, but
we don't have a relation on the downstream to apply it to, so
apply fails with an error like:

 ERROR:  42P01: relation "public.pg_temp_3005813" does not exist
 LOCATION:  RangeVarGetRelidExtended, namespace.c:413

Per the comments, handle this by simply discarding the rewrite
data. It's not pretty, but this is a "shouldn't happen" case
anyway.

Includes unit tests. Note that the changes *appear* to be replicated in
the tests, but that's becuse the DDL got replicated, and it added the
changes on the other end. Before this fix, the tests failed with
replication stopping and the DDL not getting applied.