Add hidden option --bdr-init-node to bdr_dump
And use it for logical node initialization.

Currently this option just turns on dumping of dropped column in same
way as binary upgrade does so that the definition of table on the new
node is same as on the source node including the dropped column. This is
needed because our protocol assumes same number of attributes for tables
on all nodes.

Fixes #113 and #114