Use workers instead of slots to wait for inbound connections
The slots become active before the decoding reached consistent point so
we can't be sure that every transaction after the slot became active
will be replicated. The wal sender worker shmem struct is only
initialized from the startup_cb interface of output plugin which is
called after the consistent point was found.

Fixes #161