Fix vacuumdb -a and other commands that touch all databases
Some commands want to connect to all DBs, but the client authentication was
preventing that for the supervisor DB. This didn't cause issues with xid
wraparound because autovacuum isn't affected by the client auth hook, but it
did break clients that didn't expect ERROR on connect to a DB.

Make the client auth hook ERROR into a WARNING so connections can continue.
Then check for the supervisordb in the command filter and if it's the active
database refuse to allow anything except VACUUM and ANALYZE to run.

In the process revert e19a405 which tried to fix this by changing the
command filter check into a WARNING. This allowed users to attempt to
drop/create/rename bdr_supervisordb, but didn't affect connecting to it.

Fixes #154