Add custom connstring and a default connstring added to all client connections
Define a new default connection string fragment that enables keepalives
by default.

Also add a new bdr.extra_apply_connection_options setting that users may
use to add to / override the built-in connection string. It doesn't replace
the default options, so just setting it to "" (the default) won't clear the
defaults. You have to override them with e.g. "keepalives=0". This is to stop
users cloberring default params accidentally when addding an unrelated one,
or not getting new default options added in an upgrade.

Closes #173