Add confirmed_flush column to pg_replication_slots.
This is a modified squashed commit of 3f811c2d from 9.6. Unlike the
original versions it does not modify the catalog definitions of
pg_replication_slots and its backing function
pg_get_replication_slots() since it's intended as an in-place upgrade
for 9.4bdr users.

The pg_get_replication_slots() function is modified to return the new
column as the last col. Extra cols returned from a SRF are ignored so
we can keep using the existing pg_proc definition then add our own in
an extension (like BDR) to expose the new columns.

For #186