First cut of new DDL replication
On 9.6 use the new ProcessUtility_hook based DDL replication logic instead of
event triggrers and deparse, since we don't have working ddl deparse for 9.6
yet.

Run DDL tests. They still fail due to a few behaviour differences:

* UNLOGGED tables aren't replicated (yet)
* Multi-statements may not contain DDL or COMMIT
* Functions may not perform DDL directly, only via bdr_replicate_ddl_command,
  since we can't easily tell the difference between a function body running
  something and a multi-statement.

and lots of psql formatting noise.

It has the unfixable limitation that it can affect the wrong object if it's
found earlier on the search_path on the downstream than it was on the upstream,
since we don't rewrite statements to force fully-qualified identifiers. We
should advise users to use search_path = '' as much as possible.

However, unlike the deparse approach, we can't encounter DDL we don't know how
to deparse or that we deparse incorrectly.

Also separates out DDL rep and truncation infrastruture a bit.