Replicate the ch_timeframe argument to conflict handlers
The ch_timeframe argument to bdr.bdr_create_conflict_handler was not
correctly replicated to peer nodes.

Per customer report.

Fix by refactoring the routine to use the same argument building
for both applying the change and replicating it. Apply the same
change to bdr_drop_conflict_handler.

To ensure that object names are fully qualified, set an empty
search path in the functions.