BDR 1.0.3 with upgrade compatibility for BDR 2.0
BDR 2.0 adds a number of columns to some catalog tables. Since we
use pg_dump and pg_restore as part of logical node init, and it doesn't
specify column-lists to COPY, we need to have matching columns.

So add the new columns from 2.0 to 1.0.

We can avoid this step if we update users to 2.0 on 9.4 before upgrading to 2.0
on 9.6.