Fix deparsing of CREATE SEQUENCE ... OWNED BY
Per GitHub #284, CREATE SEQUENCE failed to replicate "OWNED BY".

The relevant code (line 3942) contained this comment:

    /* we purposefully do not emit OWNED BY here */

(introduced in commit d99242f2), but no indication of the reason for doing
this. Deparsing for ALTER SEQUENCE ... OWNED BY was added in the same commit.
I'm unable to determine any issues which could be caused by emitting OWNED BY
here.

No test added as the change is not verifiable with local tests.

A test covering OWNED BY was previously added in commit 18508dc8.