Begin BDR 3.0 development
We start 3.0 from a new tree because the majority of BDR 2.0's code and
structure will simply be deleted by 3.0. Code that is re-used, like the
sequences and DDL locking support, command filter, etc, will be heavily
restructured. So there's not much point retaining a history that
starts with a giant 'git rm'.