Work around a deadlock in testing after changehub merge
Now that creating a subscription creates a slot directly in the
create func, slot creation can deadlock waiting on the create
func to commit so it can get its snapshot. But the func won't
commit until the slot is created. Only an issue when pub
and sub are on the same machine.

Work around by pre-creating the slot in tests.

See 2ndQuadrant/pglogical_internal#152