WIP: start laying down node join now that create works
Now that the catalogs functions for node creation work and we can
make local BDR nodes, it's time to add the important part, the co-ordination
of joining and subscribing between all nodes.

This is very (very) WIP, lots of stubs. Need to:

- Finish bdr_join_copy_remote_nodes
- Finish rest of funcs in bdr_join_nodegroup_sql
  (create subs, origins, slots, copy peer nodes, copy repset members,
   announce join completion, etc)
- Delay starting messaging until have nodelist and nodegroup (manager)
- Implement consensus operation handlers for (not real consensus yet, but)
  - node name reserve
  - node standby (make slots)
  - node active (make origins+subscriptions, start replaying)

However, it's mostly just directly implementing the design doc, not
complicated.