Prepare to fix a thinko in join
Every node needs a copy of each other node's interface, not just its nodes
entry, so it can talk to it. And we need to know which interface to connect to
for each node. BDR only cares about the pgl node, and pgl only stores mappings
of node->interface for subscriptions.

To do this right We'll need to look up BDR's subscriptions to find the
interface used to subscribe to each peer on the join target. But as an initial
hack we'll just require that an interface with the same name as the node exist,
and fail otherwise.

Should really allow for different interfaces facing different nodes, but no
practical way to do that yet since we discover our peers from an existing peer.
That's a TODO for later.