Perpetrate a vile hack to make join work
To get rid of this, the join guts need to move out of a SQL func
and into the manager. At that point the whole join_finish hack can
go away, as can the mess in the handle join message code.

This is now only held back by a bug in setting the dbname of the
bdr node entry, causing bad slot name generation.