Add a table-based state machine for BDR node state
Define a state journal that we'll use for node part/join state
changes, etc. It'll also be used when executing interim steps
for global consensus messages, when we have to commit something
but still need extra steps.

This should provide a more robust platform for making sure everything happens
consistently in order, in a traceable/debuggable way, and happens the same on
every node.

It's a table-based state machine, so it's not intended for fine grained state
changes within the system, only for coarse persistent state like node join
phases.

Next step is to use this to manage node join.