RM#831 Add a state journal decoding function
Add a function that allows for decoding of the state journal's state
IDs as human readable names, and decodes any state extradata in human
readable form.

A similar function will be needed for global consensus messages once
they're persisted on disk.

Use the function, with some masking for variable output, in the tests.