RM#1084 Let join pause at catchup to support logical standby
There's now a "goal" state in the state journal. When we reach the potential
stable state of STANDBY during join, we check it to see if we should proceed
further. By default nodes start up as fully active members. They can be made to
run as logical standbys by passing the new optional argument

    bdr.join_node_group(..., pause_in_standby := true)

A standby can be nudged to continue join into an active member
by the new function.

	bdr.promote_node().

This commit also adds the beginnings of maintenance of the local state in
bdr.node, which we'll need for join progress tracking.