Prevent infinite loop in apply worker on DDL lock release
When a peer node released a DDL lock, and during startup, it was
possible for BDR to enter an infinite loop in `bdr_locks_on_unlock`.
It's suspected that the cause here is a linked list with a cycle,
so introduce guards against that both on insertion and removal of
elements.

The root cause is not yet identified.

See GH#130